Moving agents files to match the api route format
